requirements.txt
hidapi
pyobjc-framework-Quartz
pyobjc-framework-ApplicationServices
rumps
